user_detail_bypthon/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
pymysql
sqlalchemy